requirements.txt
# Core dependencies
gradio>=5.5.0
pandas>=1.5.0
numpy>=1.22.0
matplotlib>=3.5.0
python-dotenv>=1.0.0


# Argilla for database
supabase==2.8.0

# LLM API dependencies
openai
google-generativeai>=0.3.0

# Optional: for development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
adalflow
